// air-statistic-app-ui/src/app/page/dashboard/search/station-search-form.interface.ts
export interface StationSearchFormModel {
  stationNumber: any;
  stationName: any;
  stationCode: any;
  stationType: any;
  internationalCode: any;
  voivodeship: any;
  address: any;
  location: any;
  locationType: any;
  indicator: any;
  measurementType: any;
  measurementStartDate: any;
  measurementEndDate: any;
}